refactor(dictionary): split into replacements/ + vocabulary/ submodules (#39)
The `dictionary` module mixed two unrelated subsystems behind one
namespace: post-transcription find/replace (replacements) and
pre-transcription Whisper prompt biasing (vocabulary). They share
the database connection and migration 0001, otherwise they're
independent.

After this PR:

  src-tauri/src/dictionary/
    mod.rs                  ← thin: re-exports + module docstring
    replacements/
      mod.rs                ← types, trait, apply_replacements()
      sqlite.rs             ← SqliteReplacementRepository
    vocabulary/
      mod.rs                ← types, trait, format_vocabulary_prompt(),
                              MAX_PROMPT_CHARS
      sqlite.rs             ← SqliteVocabularyRepository

Pure file reorganization. Public API at the `dictionary` crate
boundary is preserved via explicit named re-exports — every
existing call site (`crate::dictionary::ReplacementRule`,
`crate::dictionary::apply_replacements`,
`crate::dictionary::VocabularyTerm`, etc.) keeps resolving with
zero edits outside the module.

A contributor's first guess for "where does X live" now matches the
filesystem: replacements code under replacements/, vocabulary code
under vocabulary/. Future expansion (a regex mode for replacements,
a vocabulary import path) has a natural home.

Validation: cargo test --lib 135 pass / 1 ignored, cargo clippy
clean, cargo fmt clean. No file outside `dictionary/` modified.

Closes #39.
